requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
filetype
pdfplumber
python-docx
Pillow
pytesseract
pydub
openai-whisper
langgraph
langchain
python-dotenv
